# services/client-api/api_specification.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TaxEase Backend API",
    "description": "Comprehensive API for Canadian tax preparation and filing system with Flutter mobile app integration",
    "version": "1.0.0",
    "contact": {
      "name": "TaxEase Support",
      "email": "support@taxease.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.taxease.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "first_name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "email_verified": {
            "type": "boolean",
            "example": true
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-06T08:00:00.000000"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "User Registration",
        "description": "Register a new user account",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "first_name", "last_name", "accept_terms"],
                "properties": {
                  "email": {"type": "string", "format": "email", "example": "user@example.com"},
                  "password": {"type": "string", "minLength": 8, "example": "SecurePassword123!"},
                  "first_name": {"type": "string", "example": "John"},
                  "last_name": {"type": "string", "example": "Doe"},
                  "phone": {"type": "string", "example": "+1234567890"},
                  "accept_terms": {"type": "boolean", "example": true}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string", "example": "User registered successfully"},
                    "user": {"$ref": "#/components/schemas/User"},
                    "otp_sent": {"type": "boolean", "example": true}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email already registered or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {"type": "string", "example": "Email already registered"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User Login",
        "description": "Login with email and password to get JWT tokens",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email", "example": "user@example.com"},
                  "password": {"type": "string", "example": "SecurePassword123!"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {"type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},
                    "refresh_token": {"type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},
                    "token_type": {"type": "string", "example": "bearer"},
                    "expires_in": {"type": "integer", "example": 3600},
                    "user": {"$ref": "#/components/schemas/User"}
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {"type": "string", "example": "Invalid email or password"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}